results/rebuttal_new_baselines/run-20231112-040610_CodeT-SelfRefine-Reflexion-ZeroShot_donnemartin-system-design-oop-recipe-donnemartin-system-design-oop-eventplanner_0_20-runs_log/eventplanner/3/CodeT/app.py
from flask import Flask, request, jsonify
from dataclasses import dataclass

app = Flask(__name__)

# Mock database
DB = {}

@dataclass
class Event:
	id: int
	type: str
	date: str
	time: str
	theme: str
	color_scheme: str

@app.route('/event', methods=['POST'])
def create_event():
	data = request.get_json()
	event = Event(**data)
	DB[event.id] = event
	return jsonify(data), 201

@app.route('/event/<int:event_id>', methods=['PUT'])
def update_event(event_id):
	data = request.get_json()
	DB[event_id] = Event(**data)
	return jsonify(data), 200

if __name__ == '__main__':
	app.run(debug=True)
